+++ guard/src/rules/libyaml/mod.rs
@@ -4,3 +4,4 @@ pub mod parser;
 pub mod tag;
 pub mod error;
 pub mod loader;
+pub mod event;
